--- src/app/congrats/[resource_id]/components/index.ts
@@ -1,2 +1,0 @@
-export * from './CongratsScreen'
-export * from './RRSSButton'
