--- src/app/register/[resource_id]/components/index.ts
@@ -1,2 +1,0 @@
-export * from './RegisterFormScreen'
-export * from './SplashLoaderModal'
